feat: rewrite sd-tag documentation
