+++ packages/components/src/components/tag/tag.stories.ts
@@ -1,21 +1,18 @@
 import '../../solid-components';
 import { html } from 'lit';
 import { storybookDefaults, storybookHelpers, storybookTemplate } from '../../../scripts/storybook/helper';
-import { userEvent } from '@storybook/test';
-import { waitUntil } from '@open-wc/testing-helpers';
-import { withActions } from '@storybook/addon-actions/decorator';
 
 const { argTypes, parameters } = storybookDefaults('sd-tag');
 const { generateTemplate } = storybookTemplate('sd-tag');
 const { overrideArgs } = storybookHelpers('sd-tag');
 
 export default {
+  tags: ['!dev'],
   title: 'Components/sd-tag',
   component: 'sd-tag',
   args: overrideArgs({ type: 'slot', name: 'default', value: 'Tag' }),
   argTypes,
-  parameters: { ...parameters },
-  decorators: [withActions] as any
+  parameters: { ...parameters }
 };
 
 /**
@@ -29,173 +26,105 @@ export const Default = {
 };
 
 /**
- * The tag in all possible combinations of `selected` and `size`.
+ * The tag with the `selected` attribute.
  */
 
-export const selectedAndSize = {
-  name: 'Selected × Size',
-  parameters: { controls: { exclude: ['size', 'selected'] } },
-  render: (args: any) => {
-    return generateTemplate({
-      axis: {
-        x: { type: 'attribute', name: 'selected', values: ['false', 'true'] },
-        y: { type: 'attribute', name: 'size' }
-      },
-      args
-    });
-  }
+export const Selected = {
+  render: () => html`
+    <div class="flex gap-12">
+      <sd-tag selected>Selected</sd-tag>
+    </div>
+  `
 };
 
 /**
- * The tag in all possible combinations of `removable` and `size`.
+ * Use the `size` attribute to change the tag size. The default size is `lg`.
  */
 
-export const removableAndSize = {
-  name: 'Removable × Size',
-  parameters: { controls: { exclude: ['size', 'removable'] } },
-  render: (args: any) => {
-    return generateTemplate({
-      axis: {
-        x: { type: 'attribute', name: 'removable', values: ['false', 'true'] },
-        y: { type: 'attribute', name: 'size' }
-      },
-      args
-    });
-  }
+export const Size = {
+  render: () => html`
+    <div class="flex gap-12">
+      <sd-tag size="lg">Large</sd-tag>
+      <sd-tag size="sm">Small</sd-tag>
+    </div>
+  `
 };
 
 /**
- * Use the `disabled` attribute to disable a tag. Clicks will be suppressed until the disabled state is removed.
+ * Use the `removable` attribute to include the removability indicator.
+ */
+
+export const Removable = {
+  render: () => html`
+    <div class="flex gap-12">
+      <sd-tag removable>Removable</sd-tag>
+    </div>
+  `
+};
+
+/**
+ * Use the `removable-indicator` slot to change the removability indicator. By default a close icon is used.
+ */
+
+export const RemovableIndicatorSlot = {
+  render: () => html`
+    <div class="flex gap-12">
+      <sd-tag removable>Default</sd-tag>
+      <sd-tag removable>
+        Custom
+        <sd-icon slot="removable-indicator" library="system" name="eye" label="remove"></sd-icon>
+      </sd-tag>
+    </div>
+  `
+};
+
+/**
+ * Use the `disabled` attribute to disable a tag.
  *
  * __Hint:__ If the href attribute is set i.e. the tag is rendered as a link (`<a>`),
  * the disabled attribute is ignored, as links may not be disabled.
- * To disable the tag in this case the href attribute has to be removed as well.
+ * To disable the tag in this case the href attribute has to be removed.
  */
 
 export const Disabled = {
-  name: 'Disabled',
-  parameters: { controls: { exclude: ['selected', 'removable', 'disabled'] } },
-  render: (args: any) => {
-    return generateTemplate({
-      axis: {
-        x: { type: 'attribute', name: 'selected', values: ['false', 'true'] },
-        y: { type: 'attribute', name: 'removable', values: ['false', 'true'] }
-      },
-      constants: [
-        {
-          type: 'attribute',
-          name: 'disabled',
-          value: 'true'
-        }
-      ],
-      args
-    });
-  }
+  render: () => html`
+    <div class="flex gap-12">
+      <sd-tag disabled>Disabled</sd-tag>
+    </div>
+  `
 };
 
 /**
- * Use the `default` slot to add content to the tag.
- * Use the `removable-indicator` slot to change the removability indicator.
+ * Use the `href` attribute to render the tag as a link. By default a tag is rendered as a button.
  *
- * If you add icons to the slot, please make sure to account for accessibility by providing an alt-text.
+ * The `target` attribute can be used to specify where to open the link.
  */
 
-export const Slots = {
-  parameters: {
-    controls: { exclude: ['size', 'selected', 'removable', 'disabled'] }
-  },
-  render: (args: any) => {
-    return html`
-      ${['default', 'removable-indicator'].map(slot =>
-        generateTemplate({
-          axis: {
-            x: {
-              type: 'slot',
-              name: slot,
-              title: 'slot=...',
-              values: [
-                {
-                  title: slot,
-                  value:
-                    slot === 'default'
-                      ? `<div class="slot slot--background slot--border h-auto w-[108px] text-[8px]"></div>`
-                      : `<div slot="removable-indicator" class="slot slot--background slot--border h-auto w-4 text-[8px]"></div>`
-                }
-              ]
-            }
-          },
-          args,
-          constants: [
-            { type: 'attribute', name: 'removable', value: 'true' },
-            { type: 'attribute', name: 'size', value: 'lg' }
-          ]
-        })
-      )}
-    `;
-  }
+export const WithALink = {
+  render: () => html`
+    <div class="flex gap-12">
+      <sd-tag>Default</sd-tag>
+      <sd-tag
+        href="https://solid-design-system.fe.union-investment.de/x.x.x/storybook/?path=/docs/docs-general-introduction--docs"
+        target="_blank"
+        >Link</sd-tag
+      >
+    </div>
+  `
 };
 
 /**
- * Use the `base`, `content` and `removable-indicator` part selectors to customize the button.
+ * Use the `download` attribute to tell the browser to download the linked file as this filename. Only used when href is present.
  */
 
-export const Parts = {
-  parameters: {
-    controls: { exclude: ['size', 'selected', 'removable', 'disabled'] }
-  },
-  render: (args: any) => {
-    return generateTemplate({
-      axis: {
-        y: {
-          type: 'template',
-          name: 'sd-tag::part(...){outline: solid 2px red}',
-          values: ['base', 'content', 'removable-indicator'].map(part => {
-            return {
-              title: part,
-              value: `<style>#part-${part} sd-tag::part(${part}){outline: solid 2px red;} #part-${part} .${part}{outline: solid 2px red;}</style><div id='part-${part}'>%TEMPLATE%</div>`
-            };
-          })
-        }
-      },
-      constants: [
-        { type: 'template', name: 'width', value: '<div style="width: 300px">%TEMPLATE%</div>' },
-        {
-          type: 'slot',
-          name: 'default',
-          value: '<span class="content">Tag</span>'
-        },
-        {
-          type: 'slot',
-          name: 'removable-indicator',
-          value:
-            '<sd-icon slot="removable-indicator" library="system" name="close" label="remove" class="removable-indicator">'
-        },
-        { type: 'attribute', name: 'removable', value: 'true' },
-        { type: 'attribute', name: 'size', value: 'lg' }
-      ],
-      args
-    });
-  }
-};
-
-/**
- * sd-tags are fully accessibile via keyboard.
- */
-
-export const Mouseless = {
-  render: (args: any) => {
-    return html`<div class="mouseless">${generateTemplate({ args })}</div>`;
-  },
-
-  play: async ({ canvasElement }: { canvasElement: HTMLUnknownElement }) => {
-    const el = canvasElement.querySelector('.mouseless sd-tag');
-    await waitUntil(() => el?.shadowRoot?.querySelector('button'));
-
-    if (el?.shadowRoot) {
-      const button = el.shadowRoot.querySelector('button');
-      if (button) {
-        await userEvent.type(button, '{space}', { pointerEventsCheck: 0 });
-      }
-    }
-  }
+export const Download = {
+  render: () => html`
+    <div class="flex gap-12">
+      <sd-tag
+        href="https://github.com/solid-design-system/solid/blob/main/packages/placeholders/src/images/collaboration.jpg"
+        download
+        >Download</sd-tag
+      >
+    </div>
+  `
 };
